+++ src/main/java/com/springboot/publicplace/service/S3UploadService.java
@@ -1,35 +1,10 @@
 package com.springboot.publicplace.service;
 
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.ObjectMetadata;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 
-@Service
-@Slf4j
-@RequiredArgsConstructor
-public class S3UploadService {
-    private final AmazonS3 amazonS3;
+public interface S3UploadService {
 
-    @Value("${cloud.aws.s3.bucket}")
-    private String bucket;
-
-    public String saveFile(MultipartFile multipartFile) throws IOException {
-        String originalFilename = multipartFile.getOriginalFilename();
-
-        String fileKey = "publicplace/" + originalFilename;
-
-        ObjectMetadata metadata = new ObjectMetadata();
-        metadata.setContentLength(multipartFile.getSize());
-        metadata.setContentType(multipartFile.getContentType());
-
-        amazonS3.putObject(bucket, fileKey, multipartFile.getInputStream(), metadata);
-
-        return amazonS3.getUrl(bucket, fileKey).toString();
-    }
+    public String saveFile(MultipartFile multipartFile) throws IOException;
 }
