+++ src/main/java/com/springboot/publicplace/service/impl/S3UploadServiceImpl.java
@@ -1,0 +1,36 @@
+package com.springboot.publicplace.service.impl;
+
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.model.ObjectMetadata;
+import com.springboot.publicplace.service.S3UploadService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor
+public class S3UploadServiceImpl implements S3UploadService {
+    private final AmazonS3 amazonS3;
+
+    @Value("${cloud.aws.s3.bucket}")
+    private String bucket;
+
+    public String saveFile(MultipartFile multipartFile) throws IOException {
+        String originalFilename = multipartFile.getOriginalFilename();
+
+        String fileKey = "publicplace/" + originalFilename;
+
+        ObjectMetadata metadata = new ObjectMetadata();
+        metadata.setContentLength(multipartFile.getSize());
+        metadata.setContentType(multipartFile.getContentType());
+
+        amazonS3.putObject(bucket, fileKey, multipartFile.getInputStream(), metadata);
+
+        return amazonS3.getUrl(bucket, fileKey).toString();
+    }
+}
